worker: split the sandbox setup stage out of linux.rs
No behavior change.
